\name{addresses}
\alias{addresses}
\docType{data}
\title{addresses/data}
\description{Sample address data for geocoding}
\usage{data(addresses)}
\format{
  A data frame with 250 observations on the following 6 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{address}}{a factor with levels \code{1004 HIGH MEADOW LANE} \code{1005 GOLDEN STREET} \code{1009 WASHINGTON STREET} \code{1010 AARON SMITH DRIVE} \code{1015 SPRING AVENUE} \code{1017 HARLEY BROOK LANE} \code{1021 HEAVNER COURT} \code{1031 HAMILTON DRIVE} \code{1035 RINEHART ROAD} \code{104 PLATINUM DRIVE} \code{1045 POPLAR LANE} \code{1048 DEANS LANE} \code{1053 GNATTY CREEK ROAD} \code{1053 WINES LANE} \code{1057 COUNTY LINE ROAD} \code{1075 MEDICAL CENTER DRIVE} \code{108 LOSH LANE} \code{1080 STUART STREET} \code{1084 COUNTY LINE ROAD} \code{1087 GLEN FALLS ROAD} \code{1088 CHARMAINE LANE} \code{1088 WHISPERING PINES CIRCLE} \code{1093 DEERCOVE DRIVE} \code{1095 WYATT STREET} \code{1096 JAMES STREET} \code{1105 PATTERSON ROAD} \code{1106 GLENVIEW DRIVE} \code{1111 BROOKE STREET} \code{1116 PINE STREET} \code{1120 SCENICVIEW DRIVE} \code{1126 JARVIS STREET} \code{1134 STATION STREET} \code{1135 MOUNT TABOR} \code{1138 WASHINGTON STREET} \code{1142 JARVIS STREET} \code{115 BRENTWOOD DRIVE} \code{1158 HORSESHOE LANE} \code{1159 FARLAND AVENUE} \code{1160 BURNING MEMORY LANE} \code{1160 CARRIAGE LANE} \code{1160 ROMINES MILL ROAD} \code{1166 GIRAFFE HILL DRIVE} \code{1169 BICETOWN ROAD} \code{1179 STILES STREET} \code{1179 WORTHINGTON DRIVE} \code{118 DUNCAN AVENUE} \code{1181 BEEGHLEY STREET} \code{1185 JARVIS STREET} \code{1188 WHITETAIL LANE} \code{1191 QUINCY STREET} \code{1195 MULBERRY STREET} \code{1199 CHAPEL STREET} \code{12 FORMULA LANE} \code{1207 STONEY LONESOME ROAD} \code{1210 CLAIR STREET} \code{1215 WALT NUZUM FARM ROAD} \code{1217 HILL HAVEN DRIVE} \code{1223 SETH STREET} \code{1224 BARNES STREET} \code{123 WORTHINGTON DRIVE} \code{1235 WESTWOOD AVENUE} \code{1241 WORTHINGTON DRIVE} \code{1251 STANLEY AVENUE} \code{1252 WOODROW WAY} \code{1253 SHINN STREET} \code{1254 TERRY LANE} \code{1258 DUNCAN AVENUE} \code{1261 SOUTH STREET} \code{1272 CHARMAINE LANE} \code{1274 HINKLE DEEGAN LAKE ROAD} \code{1282 MICHIGAN AVENUE} \code{1287 LONG STREET} \code{1293 CLAIR STREET} \code{1299 STEVE HUNT ROAD} \code{1305 MICHIGAN AVENUE} \code{1307 COAL ROAD} \code{1307 FIDLER DRIVE} \code{1311 OLD DEAR LANE} \code{1316 BROOKVIEW DRIVE} \code{1316 MULBERRY LANE} \code{1317 ANMOORE ROAD} \code{1321 OAKRIDGE LANE} \code{1332 ARBUTUS DRIVE} \code{1335 SUSSEX COURT} \code{1338 RICHLAND AVENUE} \code{1339 STUART STREET} \code{1348 CHARMAINE LANE} \code{1349 FARNUM ROAD} \code{1350 MICHAEL STREET} \code{1356 HILLTOP DRIVE} \code{137 STONE LANE} \code{1371 OCALA STREET} \code{1372 ARROWOOD DRIVE} \code{1384 GRAND AVENUE} \code{1384 MORRIS STREET} \code{1394 FRANKLIN AVENUE} \code{1398 NORMA AVENUE} \code{140 EVERETTE ALLEY} \code{1406 MOORE AVENUE} \code{1410 BOONE STREET} \code{1419 MARIGOLD LANE} \code{1420 STONEYBROOK ROAD} \code{1421 ADAMS DRIVE} \code{1423 BIRCH  STREET} \code{1426 AMERICAN DRIVE} \code{1428 SPRING AVENUE} \code{1428 VILLAGE VIEW DRIVE} \code{1430 POPLAR LANE} \code{1435 STUTLER LANE} \code{1438 OAK DRIVE} \code{1443 GRAND AVENUE} \code{1445 WASHINGTON STREET} \code{1450 BUBBY DRIVE} \code{1452 WERNINGER STREET} \code{1453 WATSON LANE} \code{1465 AARON SMITH DRIVE} \code{1466 WARNER STREET} \code{1469 SAINTS ALLEY} \code{1479 POWDER HOUSE ROAD} \code{1482 TAYLOR STREET} \code{1485 DEANS LANE} \code{1485 SARDIS STA} \code{1495 HEAVENS WAY} \code{1500 PAYNE STREET} \code{1507 COLLINS STREET} \code{151 HORSESHOE LANE} \code{1510 BELL STREET} \code{1512 BURWELL HEIGHTS ROAD} \code{1513 WOODLAND TERRACE} \code{1514 TYLER AVENUE} \code{1517 RINEHART ROAD} \code{1519 WALDECK STREET} \code{152 GREY FOX FARM ROAD} \code{1520 COAL ROAD} \code{1528 WINES LANE} \code{1531 JARVISVILLE ROAD} \code{1534 CHURCH STREET} \code{1540 GIRAFFE HILL DRIVE} \code{1542 CHERRY TREE DRIVE} \code{1549 WYATT STREET} \code{155 WYATT STREET} \code{1554 GENEVA STREET} \code{1555 NEW YORK AVENUE} \code{1558 JACOBS STREET} \code{1561 CHERRY RIDGE DRIVE} \code{1562 MARGARET STREET} \code{1565 TURKEY PEN ROAD} \code{1566 ANTHONY AVENUE} \code{157 STILES STREET} \code{1575 FELOSA DRIVE} \code{1589 SARDIS STA} \code{1596 PAYNE STREET} \code{1601 SHORT STREET} \code{1612 BRENTWOOD DRIVE} \code{1613 COAL STREET} \code{1614 PHEASANT RIDGE ROAD} \code{1621 STANLEY AVENUE} \code{1623 STILES STREET} \code{1633 CHURCH STREET} \code{164 SPADAFORE DRIVE} \code{1652 LINCOLN DRIVE} \code{1656 ABIA MARTIN DRIVE} \code{1657 OLEN THOMAS DRIVE} \code{1661 WILLIS AVENUE} \code{1665 BICETOWN ROAD} \code{1671 COLLINS STREET} \code{1671 SIMPSON AVENUE} \code{1672 STONEYBROOK ROAD} \code{1681 MARYLAND AVENUE} \code{1683 MARIGOLD LANE} \code{1684 FOLEY STREET} \code{1692 AMERICAN DRIVE} \code{1694 GAMBLER LANE} \code{1694 HANOVER STREET} \code{1697 CONFEDERATE DRIVE} \code{1699 CRESTVIEW TERRACE} \code{170 HAMILTON DRIVE} \code{1704 LONG STREET} \code{1714 DRAINER AVENUE} \code{1720 TAYLOR STREET} \code{1722 BIRCH  STREET} \code{1726 WEST FORK DRIVE} \code{1733 DUNCAN AVENUE} \code{1746 CANTEBURY DRIVE} \code{1756 GERALDINE LANE} \code{1766 BROOKE STREET} \code{1767 HILLTOP DRIVE} \code{1770 PRIDE AVENUE} \code{1776 GNATTY CREEK ROAD} \code{1776 WARD ROAD} \code{1779 ELKVIEW DRIVE} \code{1781 PICKENS WAY} \code{1783 FANNIE STREET} \code{1795 LONGVIEW AVENUE} \code{18 SPRUCE DRIVE} \code{1806 AGRICULTURE LANE} \code{1809 MAPLEVIEW DRIVE} \code{1814 RENWICK DRIVE} \code{1822 WASHINGTON STREET} \code{1824 GROVE STREET} \code{1832 ANGUS ROAD} \code{1846 HORSESHOE LANE} \code{1855 OLEN THOMAS DRIVE} \code{1857 CAROLYNS CIRCLE} \code{1858 BOUNDARY STREET} \code{1865 CHURCH STREET} \code{1870 ROCKY ROAD} \code{188 GEORGE STREET} \code{1898 GARRETT STREET} \code{1899 SCENICVIEW DRIVE} \code{19 BAGWELL AVENUE} \code{19 CONFERENCE CENTER WAY} \code{190 MORGAN STREET} \code{1901 HANOVER STREET} \code{1908 SAINT MARYS AVENUE} \code{1909 ASHTON LANE} \code{1909 LINDEN AVENUE} \code{1913 MARION DRIVE} \code{1914 PHEASANT RIDGE ROAD} \code{1919 LAMBERTS BRANCH ROAD} \code{1921 GAMBLER LANE} \code{1921 WILKINSON COURT} \code{1929 LYNDON STREET} \code{193 MULBERRY LANE} \code{1937 MARION DRIVE} \code{1940 CUSTER STREET} \code{1947 PICKENS WAY} \code{195 JONES STREET} \code{1954 FRANKLIN AVENUE} \code{1956 MAPLEVIEW DRIVE} \code{1960 BICETOWN ROAD} \code{1964 WEST FORK DRIVE} \code{1976 ANMOORE ROAD} \code{1976 GODFREY ROAD} \code{1987 BAKER AVENUE} \code{1988 SMITHFIELD AVENUE} \code{1988 WORTHINGTON DRIVE} \code{1992 GNATTY CREEK ROAD} \code{20 FLORENCE STREET} \code{2016 BARNES STREET} \code{2021 WHITETAIL LANE} \code{2022 SAINT MARYS AVENUE} \code{2023 SUNBURST DRIVE} \code{2025 PLATINUM DRIVE} \code{203 POCO MAS DRIVE} \code{2042 CHARMAINE LANE} \code{2042 TETRICK ROAD} \code{2043 STONEY LONESOME ROAD} \code{2046 MARSHVILLE ROAD} \code{2048 CAMERON ROAD} \code{2052 FARLAND AVENUE} \code{2057 CAROLINA AVENUE} \code{206 STANLEY AVENUE} \code{2062 MONROE STREET} \code{2066 JAMES AVENUE} \code{2076 JODY ROAD} \code{2076 WAYBACK LANE} \code{208 COAL STREET} \code{2080 JARVIS STREET} \code{2082 GOLDEN RIDGE ROAD} \code{2083 CITYVIEW DRIVE} \code{2086 FORMULA LANE} \code{2092 COAL ROAD} \code{2094 MARGARET STREET} \code{2095 ERSEL STREET} \code{2096 PARRISH AVENUE} \code{2099 BADGER POND LANE} \code{2101 FLORENCE STREET} \code{2107 STOUT STREET} \code{2108 BEECHWOOD DRIVE} \code{211 BEN STREET} \code{2110 FRESHOUR CIRCLE} \code{2111 TRAVIS STREET} \code{2113 HINKLE DEEGAN LAKE ROAD} \code{2117 SPIRIT DRIVE} \code{2118 SHINN STREET} \code{2122 GEORGE STREET} \code{2126 PAYNE STREET} \code{2138 COLLINS STREET} \code{2141 ROMINES MILL ROAD} \code{2146 QUINCY STREET} \code{2148 FANNIE STREET} \code{2151 MICHIGAN AVENUE} \code{2156 TODDS LANE} \code{2157 FIDLER DRIVE} \code{2158 POPLAR LANE} \code{2170 ELKVIEW DRIVE} \code{2177 SOUTH STREET} \code{2180 ADAMSVILLE ROAD} \code{2183 OAK DRIVE} \code{2199 MULBERRY LANE} \code{2199 WEST VIRGINIA AVENUE} \code{2204 LAKE FOREST DRIVE} \code{2208 CHESTNUT STREET} \code{221 HARLEY BROOK LANE} \code{2214 JARVIS STREET} \code{2220 ELM DRIVE} \code{2228 BURWELL HEIGHTS ROAD} \code{2229 ROSEWOOD LANE} \code{2236 JOES ROAD} \code{224 PINE STREET} \code{2240 RHAPSODY STREET} \code{2241 JARVISVILLE ROAD} \code{2245 ASH STREET} \code{2252 BROOKE STREET} \code{2253 TETRICK ROAD} \code{2255 SHORT STREET} \code{2256 POE ROAD} \code{2266 CARRIAGE LANE} \code{2273 OAK STREET} \code{2278 FIELDCREST ROAD} \code{2280 DEANS LANE} \code{2283 PARRISH AVENUE} \code{2288 LUNETTA STREET} \code{2290 HALL PLACE} \code{2291 TIPPLE ROAD} \code{23 MULBERRY STREET} \code{2304 MOORE AVENUE} \code{2305 SMITHFIELD AVENUE} \code{2314 PALLET STREET} \code{2316 BADGER POND LANE} \code{2317 WESTWOOD AVENUE} \code{2320 ABIA MARTIN DRIVE} \code{2323 CHAPEL STREET} \code{2330 ASHTON LANE} \code{2330 HEMLOCK LANE} \code{2333 AARON SMITH DRIVE} \code{2336 MULBERRY STREET} \code{234 MULBERRY STREET} \code{234 PICKENS WAY} \code{2340 GORE STREET} \code{2351 RICHLAND AVENUE} \code{2357 FANCHER DRIVE} \code{237 BENEDUM DRIVE} \code{2370 LUCKY DUCK DRIVE} \code{2375 TETRICK ROAD} \code{2380 VALLEY DRIVE} \code{2390 FLORENCE STREET} \code{2391 FORMULA LANE} \code{2404 FORMULA LANE} \code{2406 PINE STREET} \code{2419 SOUTH STREET} \code{2424 GNATTY CREEK ROAD} \code{2426 BICETOWN ROAD} \code{2437 SMALL STREET} \code{2439 LUCKY DUCK DRIVE} \code{2440 PARRISH AVENUE} \code{2445 ALPHA AVENUE} \code{2450 WASHINGTON STREET} \code{2451 WINES LANE} \code{2452 MEDICAL CENTER DRIVE} \code{2453 LEO STREET} \code{2457 POCO MAS DRIVE} \code{2459 BIRD SPRING LANE} \code{2459 WASHINGTON STREET} \code{2461 COLONIAL DRIVE} \code{2462 FROSTY LANE} \code{2468 CHANDLER HOLLOW ROAD} \code{2468 LAMBERTS BRANCH ROAD} \code{2468 LINDEN AVENUE} \code{247 NUZUM COURT} \code{2476 BASTIN DRIVE} \code{2478 RICHLAND AVENUE} \code{248 BROOKE STREET} \code{2485 BIRCH  STREET} \code{2485 SPRING AVENUE} \code{2489 FANNIE STREET} \code{2496 DUNCAN AVENUE} \code{2499 FROSTY LANE} \code{2500 STONEY LANE} \code{2511 SPRUCE DRIVE} \code{2514 GEORGE STREET} \code{2521 WYATT STREET} \code{2523 CONFERENCE CENTER WAY} \code{2524 ARROWOOD DRIVE} \code{2526 LYNDON STREET} \code{2527 DRAINER AVENUE} \code{2530 ELKVIEW DRIVE} \code{2542 STONEY LANE} \code{2559 BAKER AVENUE} \code{2559 LONG STREET} \code{2564 GOLDEN RIDGE ROAD} \code{2564 MEDICAL CENTER DRIVE} \code{2565 CAROLYNS CIRCLE} \code{2566 BRIERCLIFF ROAD} \code{2571 VALLEY DRIVE} \code{2572 SUNBURST DRIVE} \code{2573 MORRIS STREET} \code{2578 PHEASANT RIDGE ROAD} \code{2579 GOLDEN RIDGE ROAD} \code{2587 WINES LANE} \code{2588 BARNES STREET} \code{2590 FANCHER DRIVE} \code{2592 CINNAMON LANE} \code{2593 DEANS LANE} \code{2593 TIPPLE ROAD} \code{2598 WASHINGTON STREET} \code{2611 GRAND AVENUE} \code{2619 ADAMSVILLE ROAD} \code{2624 CONFEDERATE DRIVE} \code{2626 FARLAND AVENUE} \code{2628 OAK DRIVE} \code{2638 MARIGOLD LANE} \code{2640 ANMOORE ROAD} \code{2643 AMERICAN DRIVE} \code{2645 BADGER POND LANE} \code{2647 FRANKLIN AVENUE} \code{2648 FLORENCE STREET} \code{2649 ERSEL STREET} \code{265 POINTE LANE} \code{2658 BAKER AVENUE} \code{2658 WATSON LANE} \code{2661 ERSEL STREET} \code{2662 CUSTER STREET} \code{2663 ARBUTUS DRIVE} \code{267 REDBUD DRIVE} \code{2672 BOONE STREET} \code{2676 ADAMSVILLE ROAD} \code{2684 BRIARHILL LANE} \code{2689 TETRICK ROAD} \code{2704 GAMBLER LANE} \code{2705 CUSTER STREET} \code{2705 FRANKLIN AVENUE} \code{2705 SMITHFIELD AVENUE} \code{2707 HIGH MEADOW LANE} \code{2708 GRANT STREET} \code{2710 MEDICAL CENTER DRIVE} \code{2711 ANMOORE ROAD} \code{2714 ALEXANDER DRIVE} \code{2715 QUINCY STREET} \code{2716 WORTHINGTON DRIVE} \code{2720 HIGH MEADOW LANE} \code{2723 HARLEY BROOK LANE} \code{2724 SETH STREET} \code{2732 ANGUS ROAD} \code{2732 BERKLEY STREET} \code{2737 HIDDEN VALLEY ROAD} \code{2742 STILES STREET} \code{2753 BOONE STREET} \code{2753 PINE STREET} \code{2754 CHARMAINE LANE} \code{2757 SCOTT STREET} \code{276 FRANKLIN AVENUE} \code{2763 LOVING ACRES ROAD} \code{2766 BOTTOM LANE} \code{277 DEERCOVE DRIVE} \code{2773 CANDLELIGHT DRIVE} \code{2773 HOLT STREET} \code{2773 MCDONALD AVENUE} \code{2780 PLATINUM DRIVE} \code{2790 MULBERRY STREET} \code{2791 FROSTY LANE} \code{2792 BERKLEY STREET} \code{2796 LYNDON STREET} \code{2801 OAK DRIVE} \code{2804 OLEN THOMAS DRIVE} \code{2817 NORMA AVENUE} \code{2819 WEST FORK DRIVE} \code{2821 POCO MAS DRIVE} \code{2825 TAYLOR STREET} \code{283 WOODROW WAY} \code{2834 CIRCLE DRIVE} \code{2838 COLONIAL DRIVE} \code{2839 FRESHOUR CIRCLE} \code{284 GOLDEN STREET} \code{284 MORRIS STREET} \code{284 TODDS LANE} \code{2842 GOLDEN STREET} \code{2845 ADAMS DRIVE} \code{2846 HEAVENS WAY} \code{2851 WINES LANE} \code{2856 CLIFFSIDE DRIVE} \code{2858 WALDECK STREET} \code{286 STONEYBROOK ROAD} \code{286 WEST VIRGINIA AVENUE} \code{2860 OAK DRIVE} \code{2870 BELL STREET} \code{2872 BEEGHLEY STREET} \code{29 LUCKY DUCK DRIVE} \code{290 NORMA AVENUE} \code{290 RHAPSODY STREET} \code{2905 VILLAGE VIEW DRIVE} \code{2908 COLLINS STREET} \code{2908 ROANE AVENUE} \code{2909 TAYLOR STREET} \code{292 CHURCH STREET} \code{2920 CLIFFSIDE DRIVE} \code{2927 BRANNON AVENUE} \code{2927 ERSEL STREET} \code{293 ROSEWOOD LANE} \code{293 WEST FORK DRIVE} \code{2938 WHITE AVENUE} \code{294 SAINTS ALLEY} \code{2945 ABIA MARTIN DRIVE} \code{2949 MULBERRY STREET} \code{2966 CHARLA LANE} \code{2967 BOTTOM LANE} \code{2970 SETTLERS LANE} \code{2974 TAYLOR STREET} \code{2989 MARSHVILLE ROAD} \code{2989 PALLET STREET} \code{2992 MICHIGAN AVENUE} \code{2993 SOUTH STREET} \code{2996 SPADAFORE DRIVE} \code{2998 WHITE AVENUE} \code{3 BROWNING LANE} \code{3 COLLINS STREET} \code{3001 TRAILS END ROAD} \code{3012 BOUNDARY STREET} \code{3017 AGRICULTURE LANE} \code{302 FORMULA LANE} \code{302 GERALDINE LANE} \code{3025 SAINT JAMES DRIVE} \code{3036 CHAPEL STREET} \code{3042 BEECHWOOD DRIVE} \code{3047 STUTLER LANE} \code{3060 CUSTER STREET} \code{3066 ALFRED DRIVE} \code{3082 AARON SMITH DRIVE} \code{3083 WALT NUZUM FARM ROAD} \code{3084 PARRISH AVENUE} \code{3085 BOTTOM LANE} \code{3085 HILLTOP DRIVE} \code{3091 ALFRED DRIVE} \code{3091 CONFEDERATE DRIVE} \code{3094 DUNCAN AVENUE} \code{3097 SAINT JAMES DRIVE} \code{3101 STUART STREET} \code{3101 WORTHINGTON DRIVE} \code{3105 HILLTOP DRIVE} \code{3107 AARON SMITH DRIVE} \code{3109 NEW YORK AVENUE} \code{3112 TRAVIS STREET} \code{3113 BIRD SPRING LANE} \code{3116 STATION STREET} \code{3123 CONFEDERATE DRIVE} \code{3126 STONEY LONESOME ROAD} \code{3130 REDBUD DRIVE} \code{3134 GROVE STREET} \code{3136 HAMILTON DRIVE} \code{3138 STUTLER LANE} \code{3142 ELM DRIVE} \code{3145 SUNBURST DRIVE} \code{3146 WOODSIDE CIRCLE} \code{3152 CANDLELIGHT DRIVE} \code{3158 FRANK AVENUE} \code{3159 CIRCLE DRIVE} \code{3159 RIDENOUR STREET} \code{3163 COLLINS STREET} \code{3169 NEW YORK AVENUE} \code{3175 OLIVER STREET} \code{3183 HOPE STREET} \code{3187 BAGWELL AVENUE} \code{3191 MOUNT TABOR} \code{3193 CANDLELIGHT DRIVE} \code{3194 LAUREL LANE} \code{3194 MARYLAND AVENUE} \code{32 JACOBS STREET} \code{32 STONEY LANE} \code{3204 LAMBERTS BRANCH ROAD} \code{3206 JAMES STREET} \code{3210 CAMERON ROAD} \code{3215 GRAND AVENUE} \code{3218 COAL STREET} \code{322 MOORE AVENUE} \code{3226 STONEYBROOK ROAD} \code{3233 DRAINER AVENUE} \code{3234 RHAPSODY STREET} \code{3238 WHITE AVENUE} \code{3242 FANCHER DRIVE} \code{3251 TIPPLE ROAD} \code{3252 FARLAND AVENUE} \code{3254 JONES STREET} \code{3254 PLEASANT HILL ROAD} \code{3262 COAL ROAD} \code{3263 COAL STREET} \code{3263 FRANKLIN AVENUE} \code{3264 OAK STREET} \code{3268 WHISPERING PINES CIRCLE} \code{3269 ADAMSVILLE ROAD} \code{327 STANLEY AVENUE} \code{3273 MORRIS STREET} \code{3273 TRAVIS STREET} \code{3274 OAKWOOD AVENUE} \code{3281 HARDESTY STREET} \code{3284 CLINTON STREET} \code{3288 BRENTWOOD DRIVE} \code{33 SUNBURST DRIVE} \code{3303 COOLIDGE STREET} \code{3304 CHARLA LANE} \code{3304 FOREST AVENUE} \code{3304 GODFREY ROAD} \code{3304 RIDENOUR STREET} \code{3304 WASHINGTON STREET} \code{3305 AARON SMITH DRIVE} \code{3309 PRIDE AVENUE} \code{3311 SPIRIT DRIVE} \code{3318 STILES STREET} \code{3320 HANOVER STREET} \code{3329 COAL ROAD} \code{3329 RIDENOUR STREET} \code{3343 WALDECK STREET} \code{3344 HEAVNER COURT} \code{3347 MULBERRY LANE} \code{3348 BEECHWOOD DRIVE} \code{3348 GEORGE STREET} \code{3353 LADY BUG DRIVE} \code{3354 DANCING DOVE LANE} \code{3358 RHAPSODY STREET} \code{3359 WARD ROAD} \code{3367 LADY BUG DRIVE} \code{3368 HINKLE DEEGAN LAKE ROAD} \code{3371 MY DRIVE} \code{3372 MARYLAND AVENUE} \code{3374 MARION DRIVE} \code{3376 CHERRY RIDGE DRIVE} \code{3378 SHINN STREET} \code{3388 MCDONALD AVENUE} \code{3393 RINEHART ROAD} \code{3396 WHISPERING PINES CIRCLE} \code{3398 WHISPERING PINES CIRCLE} \code{3399 BASTIN DRIVE} \code{34 KENWOOD PLACE} \code{3420 OAKDALE AVENUE} \code{3421 ROCKY ROAD} \code{3432 ERSEL STREET} \code{3443 PATTERSON ROAD} \code{345 TIPPLE ROAD} \code{3457 ANMOORE ROAD} \code{3457 BURWELL HEIGHTS ROAD} \code{3458 CINNAMON LANE} \code{3470 SHINN AVENUE} \code{3474 MARSHVILLE ROAD} \code{3474 PHEASANT RIDGE ROAD} \code{349 WEEKLEY STREET} \code{3492 SUNBURST DRIVE} \code{3492 WHITE AVENUE} \code{3495 GRAND AVENUE} \code{3496 TYLER AVENUE} \code{35 CANDLELIGHT DRIVE} \code{35 PLAINFIELD AVENUE} \code{3510 POINTE LANE} \code{3514 EVERETTE ALLEY} \code{3520 MORNINGVIEW LANE} \code{3523 RIDENOUR STREET} \code{353 BRIERCLIFF ROAD} \code{3534 TRAVIS STREET} \code{3536 EVERETTE ALLEY} \code{3552 BASTIN DRIVE} \code{3559 PHEASANT RIDGE ROAD} \code{3560 WHISPERING PINES CIRCLE} \code{3571 HALL PLACE} \code{3574 HIGH MEADOW LANE} \code{3576 ARBUTUS DRIVE} \code{358 MY DRIVE} \code{3583 GOLDEN RIDGE ROAD} \code{3583 SPRUCE DRIVE} \code{3584 OCALA STREET} \code{3585 ARBUTUS DRIVE} \code{3585 BRIERCLIFF ROAD} \code{3589 BASTIN DRIVE} \code{3594 STEVE HUNT ROAD} \code{36 FILBERT STREET} \code{36 STOUT STREET} \code{3604 FRANKLIN AVENUE} \code{361 CALDWELL ROAD} \code{3614 CHURCH STREET} \code{3617 POCO MAS DRIVE} \code{3620 SETH STREET} \code{3627 ORPHAN ROAD} \code{3628 CUSTER STREET} \code{3629 OLIVER STREET} \code{3638 COAL ROAD} \code{3638 HIDDENVIEW DRIVE} \code{3646 STONEY LANE} \code{3647 ASH STREET} \code{365 BENEDUM DRIVE} \code{365 WILKINSON COURT} \code{3650 PRIDE AVENUE} \code{3654 LIBERTY STREET} \code{3659 JEROME AVENUE} \code{366 CARRIAGE LANE} \code{3666 CHAPEL STREET} \code{367 MARSHVILLE ROAD} \code{3670 TAYLOR STREET} \code{3672 CAROLYNS CIRCLE} \code{3672 WORTHINGTON DRIVE} \code{3681 SPRUCE DRIVE} \code{3681 WOODROW WAY} \code{3686 HENRY FORD AVENUE} \code{3688 FRANK AVENUE} \code{3692 FOLEY STREET} \code{3693 SOUTHERN STREET} \code{3696 ARROWOOD DRIVE} \code{3696 SOUTHERN STREET} \code{3698 MARION DRIVE} \code{370 BIRD SPRING LANE} \code{3702 COAL ROAD} \code{3703 CHESTNUT STREET} \code{3704 COUNTY LINE ROAD} \code{3706 TETRICK ROAD} \code{3711 SMITHFIELD AVENUE} \code{3713 FRANKLEE LANE} \code{3714 BEECHWOOD DRIVE} \code{3718 HALL PLACE} \code{3720 TODDS LANE} \code{3722 HEAVENS WAY} \code{3728 SWICK HILL STREET} \code{3729 TURKEY PEN ROAD} \code{3735 CARRIAGE LANE} \code{3737 MARION DRIVE} \code{3738 MY DRIVE} \code{3740 BERNARDO STREET} \code{3740 WEST VIRGINIA AVENUE} \code{3747 BRANNON AVENUE} \code{3747 CHURCH STREET} \code{3748 VALLEY LANE} \code{375 FIELDCREST ROAD} \code{3755 QUINCY STREET} \code{3756 CAMERON ROAD} \code{3758 WERNINGER STREET} \code{3760 ARBUTUS DRIVE} \code{3760 BOTTOM LANE} \code{3772 FLORENCE STREET} \code{3778 JARVIS STREET} \code{3780 WOODROW WAY} \code{3784 CIRCLE DRIVE} \code{3794 MARION DRIVE} \code{3800 BADGER POND LANE} \code{3802 CONFERENCE CENTER WAY} \code{3809 WEST VIRGINIA AVENUE} \code{3810 WHITETAIL LANE} \code{3811 CANTEBURY DRIVE} \code{3811 SAINT MARYS AVENUE} \code{3812 SAINT JAMES DRIVE} \code{3823 BAKER AVENUE} \code{3824 SETTLERS LANE} \code{3826 WORTHINGTON DRIVE} \code{3830 LUCKY DUCK DRIVE} \code{3830 ROSEWOOD LANE} \code{3832 CHARLA LANE} \code{3833 SCOTT STREET} \code{3835 MY DRIVE} \code{3838 BRANNON AVENUE} \code{3841 OAKDALE AVENUE} \code{3845 CIRCLE DRIVE} \code{3846 FELOSA DRIVE} \code{385 CHURCH STREET} \code{3876 GENEVA STREET} \code{3876 LIBERTY STREET} \code{388 BUBBY DRIVE} \code{3882 PATTERSON ROAD} \code{3885 BIRD SPRING LANE} \code{3893 WARD ROAD} \code{3896 PRIDE AVENUE} \code{3898 LONGVIEW AVENUE} \code{3899 MOUNT TABOR} \code{3907 BARNES STREET} \code{3907 WILSON AVENUE} \code{391 TIPPLE ROAD} \code{3910 DUNCAN AVENUE} \code{3910 SPADAFORE DRIVE} \code{3916 ROCKY ROAD} \code{3916 VALLEY LANE} \code{3922 GOLDEN RIDGE ROAD} \code{3925 FRANKLEE LANE} \code{3928 WARD ROAD} \code{3933 ARBUTUS DRIVE} \code{3936 SOUTHERN STREET} \code{3936 TERRY LANE} \code{394 SUSSEX COURT} \code{3941 REDBUD DRIVE} \code{3947 BROOKVIEW DRIVE} \code{3949 ANTHONY AVENUE} \code{3956 LONG STREET} \code{3957 WILSON AVENUE} \code{3961 PLATINUM DRIVE} \code{3967 MOUNT TABOR} \code{3976 SPRING AVENUE} \code{3977 POE ROAD} \code{3978 OLD DEAR LANE} \code{3989 STONEY LANE} \code{399 LINCOLN DRIVE} \code{3990 CHANDLER HOLLOW ROAD} \code{3994 ARBUTUS DRIVE} \code{4000 HILL HAVEN DRIVE} \code{4008 BARNES STREET} \code{4008 WINES LANE} \code{4009 LAUREL LANE} \code{4015 JARVIS STREET} \code{4020 VALLEY LANE} \code{4029 WILLIS AVENUE} \code{4055 BICETOWN ROAD} \code{4055 MORGAN STREET} \code{4056 JARVISVILLE ROAD} \code{4057 HORSESHOE LANE} \code{4059 HIDDEN VALLEY ROAD} \code{4059 JEROME AVENUE} \code{4061 ROSEMONT AVENUE} \code{4062 SHORT STREET} \code{4064 PLAINFIELD AVENUE} \code{4064 SARDIS STA} \code{4067 CHESTNUT STREET} \code{4070 HEMLOCK LANE} \code{4077 MARIGOLD LANE} \code{4081 LONGVIEW AVENUE} \code{4083 BAGWELL AVENUE} \code{4087 LOST CREEK ROAD} \code{4089 WEST VIRGINIA AVENUE} \code{4098 COUNTY LINE ROAD} \code{4102 JAMES STREET} \code{4104 ANMOORE ROAD} \code{4104 BENEDUM DRIVE} \code{4108 EVERETTE ALLEY} \code{4116 EMMA STREET} \code{4118 JODY ROAD} \code{4121 HENRY FORD AVENUE} \code{4130 STONEY LONESOME ROAD} \code{4135 CLIFFSIDE DRIVE} \code{4137 WESTWOOD AVENUE} \code{4140 BROOKE STREET} \code{4141 BRENTWOOD DRIVE} \code{4154 BICETOWN ROAD} \code{4155 MULBERRY STREET} \code{4159 GLEN FALLS ROAD} \code{4160 JODY ROAD} \code{4163 CAMDEN PLACE} \code{4163 KENWOOD PLACE} \code{4174 JACOBS STREET} \code{4177 GORE STREET} \code{418 JAMES STREET} \code{4183 FRANK AVENUE} \code{4184 CHARLA LANE} \code{4186 JACOBS STREET} \code{4205 GNATTY CREEK ROAD} \code{4206 CHARMAINE LANE} \code{421 CANDLELIGHT DRIVE} \code{4216 MONROE AVENUE} \code{4220 FIELDCREST ROAD} \code{4221 TURKEY PEN ROAD} \code{4223 JARVISVILLE ROAD} \code{4226 CLAIR STREET} \code{423 LOST CREEK ROAD} \code{4238 CONFEDERATE DRIVE} \code{4238 WILLIS AVENUE} \code{4241 LINDEN AVENUE} \code{4243 EVERETTE ALLEY} \code{4268 TIMBER OAK DRIVE} \code{4269 COAL ROAD} \code{4272 MARYLAND AVENUE} \code{4279 GEORGE STREET} \code{4284 BOTTOM LANE} \code{4284 KENWOOD PLACE} \code{4295 STAR TREK DRIVE} \code{4301 MAPLEVIEW DRIVE} \code{4302 CARRIAGE LANE} \code{4303 WALT NUZUM FARM ROAD} \code{4305 BOONE STREET} \code{4306 SHINN STREET} \code{4313 PICKENS WAY} \code{4313 STOUT STREET} \code{4314 BEECHWOOD DRIVE} \code{4317 QUINCY STREET} \code{4324 CHERRY TREE DRIVE} \code{4342 ANTHONY AVENUE} \code{4346 KENWOOD PLACE} \code{4352 BEECHWOOD DRIVE} \code{4356 WARD ROAD} \code{4358 CHARLA LANE} \code{4366 BURNING MEMORY LANE} \code{4373 CONFERENCE CENTER WAY} \code{4380 QUINCY STREET} \code{439 STONE LANE} \code{4392 MCDONALD AVENUE} \code{4393 GIRAFFE HILL DRIVE} \code{4395 ABIA MARTIN DRIVE} \code{4395 GLENVIEW DRIVE} \code{4397 PLATINUM DRIVE} \code{4399 OLEN THOMAS DRIVE} \code{44 ASHTON LANE} \code{440 CANTEBURY DRIVE} \code{4400 CHARLA LANE} \code{4401 ELM DRIVE} \code{4401 FRANK AVENUE} \code{441 ROCKY ROAD} \code{4412 HEAVNER COURT} \code{4415 ARROWOOD DRIVE} \code{4417 BAGWELL AVENUE} \code{4418 FREDERICK STREET} \code{4419 SIMONS HOLLOW ROAD} \code{4424 BARNES STREET} \code{4425 FANCHER DRIVE} \code{4430 BUBBY DRIVE} \code{4434 SAINT MARYS AVENUE} \code{4434 TRAILS END ROAD} \code{4435 THRASH TRAIL} \code{4441 BOONE STREET} \code{4446 LYNDON STREET} \code{4452 RINEHART ROAD} \code{4463 GNATTY CREEK ROAD} \code{4469 VIRGIL STREET} \code{4470 ASH STREET} \code{4473 MORNINGVIEW LANE} \code{4479 COLONIAL DRIVE} \code{4485 WESTWOOD AVENUE} \code{4488 FLORENCE STREET} \code{449 VALLEY DRIVE} \code{4492 BENEDUM DRIVE} \code{4493 CIRCLE DRIVE} \code{4493 HIDDEN VALLEY ROAD} \code{4494 MULBERRY LANE} \code{4495 POINTE LANE} \code{4498 HIGH MEADOW LANE} \code{450 ORPHAN ROAD} \code{4504 BAGWELL AVENUE} \code{4505 BROOKVIEW DRIVE} \code{4509 HOFFMAN AVENUE} \code{4510 ALPHA AVENUE} \code{4522 STONEY LANE} \code{4525 MARSHVILLE ROAD} \code{4526 SMITHFIELD AVENUE} \code{4529 TRAVIS STREET} \code{4530 GOLDEN STREET} \code{4537 LYNN OGDEN LANE} \code{4544 BARNES STREET} \code{4549 CLAIR STREET} \code{4555 OLD DEAR LANE} \code{4557 ROMINES MILL ROAD} \code{4558 FARLAND AVENUE} \code{4560 RICHLAND AVENUE} \code{4563 ALFRED DRIVE} \code{4568 BENEDUM DRIVE} \code{4570 FORMULA LANE} \code{4575 HIDE A WAY ROAD} \code{4578 COLLINS STREET} \code{4590 SHORT STREET} \code{4597 TREE TOP LANE} \code{46 PATTERSON ROAD} \code{460 SHINN STREET} \code{4601 RIDENOUR STREET} \code{4607 ALPHA AVENUE} \code{4617 SARDIS STA} \code{4629 GARRETT STREET} \code{4635 BOUNDARY STREET} \code{4636 BOONE STREET} \code{4636 WILSON AVENUE} \code{464 WOODSIDE CIRCLE} \code{4640 MILL STREET} \code{4643 ALPHA AVENUE} \code{4643 GROVE STREET} \code{4648 TYLER AVENUE} \code{4654 SMALL STREET} \code{4659 AARON SMITH DRIVE} \code{4663 HINKLE DEEGAN LAKE ROAD} \code{4664 PHEASANT RIDGE ROAD} \code{4670 POINTE LANE} \code{4674 SIMONS HOLLOW ROAD} \code{4691 FRANKLEE LANE} \code{4694 BURNING MEMORY LANE} \code{4695 DEANS LANE} \code{4700 RHAPSODY STREET} \code{4703 OLIVER STREET} \code{4707 SPRUCE DRIVE} \code{4712 STUART STREET} \code{4717 HILLTOP DRIVE} \code{4722 CHANDLER HOLLOW ROAD} \code{4723 LAKE FOREST DRIVE} \code{4725 BOUNDARY STREET} \code{4728 CANDLELIGHT DRIVE} \code{4728 SPRUCE DRIVE} \code{473 ELKVIEW DRIVE} \code{4734 BROOKVIEW DRIVE} \code{4738 CIRCLE DRIVE} \code{4742 WHITETAIL LANE} \code{4744 FILBERT STREET} \code{4745 SOUTHERN STREET} \code{4749 SYCAMORE CIRCLE} \code{4749 VIRGIL STREET} \code{4753 CARRIAGE LANE} \code{4754 VIRGIL STREET} \code{4758 RICHLAND AVENUE} \code{4759 ALFRED DRIVE} \code{4759 MICHAEL STREET} \code{4759 SUNDOWN LANE} \code{4775 LYNDON STREET} \code{4783 GORE STREET} \code{4787 WAKEFIELD STREET} \code{4790 CHAPEL STREET} \code{4796 CARRIAGE LANE} \code{4804 REDBUD DRIVE} \code{4804 SMITHFIELD AVENUE} \code{4808 TIMBER OAK DRIVE} \code{481 CHARDONNAY DRIVE} \code{4810 LINCOLN DRIVE} \code{4815 BOUNDARY STREET} \code{4822 MEDICAL CENTER DRIVE} \code{483 VALLEY DRIVE} \code{4833 ELKVIEW DRIVE} \code{4839 NUZUM COURT} \code{4846 THRASH TRAIL} \code{4850 LYNN OGDEN LANE} \code{4861 CLOUSSON ROAD} \code{4873 MARGARET STREET} \code{489 MORGAN STREET} \code{4890 TERRY LANE} \code{4893 OAK DRIVE} \code{4902 NUZUM COURT} \code{4904 HANOVER STREET} \code{4914 QUINCY STREET} \code{4915 JARVISVILLE ROAD} \code{4923 MICHIGAN AVENUE} \code{4924 COLONIAL DRIVE} \code{4933 WALDECK STREET} \code{4939 LINDEN AVENUE} \code{4944 GOLDEN STREET} \code{4949 GROVE STREET} \code{4956 DUNCAN AVENUE} \code{496 RHAPSODY STREET} \code{4961 FOREST AVENUE} \code{4972 OLEN THOMAS DRIVE} \code{4987 SPRING AVENUE} \code{4997 BADGER POND LANE} \code{4998 BENEDUM DRIVE} \code{4998 BRIERCLIFF ROAD} \code{4999 PATTERSON ROAD} \code{500 TAYLOR STREET} \code{502 ELKVIEW DRIVE} \code{504 JARVISVILLE ROAD} \code{505 BIRD SPRING LANE} \code{509 JAMES STREET} \code{513 STUART STREET} \code{522 TERRY LANE} \code{524 OLEN THOMAS DRIVE} \code{535 PLATINUM DRIVE} \code{537 LINDEN AVENUE} \code{540 GOLDEN RIDGE ROAD} \code{540 MY DRIVE} \code{549 MCDONALD AVENUE} \code{553 OAKRIDGE LANE} \code{557 LINDEN AVENUE} \code{558 RENWICK DRIVE} \code{566 EMMA STREET} \code{567 CANTEBURY DRIVE} \code{568 CLAIR STREET} \code{574 WOODSIDE CIRCLE} \code{575 LAKE FOREST DRIVE} \code{576 TREE TOP LANE} \code{58 HIGH MEADOW LANE} \code{580 ALPHA AVENUE} \code{586 QUINCY STREET} \code{587 BERNARDO STREET} \code{587 BIRCH  STREET} \code{587 CONFEDERATE DRIVE} \code{588 QUINCY STREET} \code{589 HEMLOCK LANE} \code{590 STEVE HUNT ROAD} \code{591 FRANKLEE LANE} \code{592 SYCAMORE CIRCLE} \code{593 AGRICULTURE LANE} \code{60 DANCING DOVE LANE} \code{61 TIPPLE ROAD} \code{611 GOLDEN STREET} \code{623 CRESTVIEW TERRACE} \code{623 WESTWOOD AVENUE} \code{627 POCO MAS DRIVE} \code{641 ADAMSVILLE ROAD} \code{645 WEST FORK DRIVE} \code{656 ELKVIEW DRIVE} \code{66 CHERRY TREE DRIVE} \code{66 FANNIE STREET} \code{66 JAMES STREET} \code{674 ETHELS LANE} \code{679 PATTERSON STREET} \code{683 WOODSIDE CIRCLE} \code{688 GANDY STREET} \code{69 RENWICK DRIVE} \code{691 JAMES AVENUE} \code{708 COAL STREET} \code{711 FELOSA DRIVE} \code{719 WYATT STREET} \code{725 HANOVER STREET} \code{726 SETH STREET} \code{727 MARSHVILLE ROAD} \code{728 FRANKLEE LANE} \code{734 BARNES STREET} \code{734 BIRD SPRING LANE} \code{735 ELM DRIVE} \code{736 STILES STREET} \code{738 CAROLYNS CIRCLE} \code{740 WILLIS AVENUE} \code{741 WALDECK STREET} \code{742 BEN STREET} \code{742 TETRICK ROAD} \code{743 GLEN FALLS ROAD} \code{750 PRIDE AVENUE} \code{754 BEECHWOOD DRIVE} \code{757 TYLER AVENUE} \code{768 BEN STREET} \code{770 BUCKHANNAN AVENUE} \code{776 CHERRY RIDGE DRIVE} \code{777 NUZUM COURT} \code{782 LAKE FOREST DRIVE} \code{783 WHISPERING PINES CIRCLE} \code{787 LYNN OGDEN LANE} \code{790 KENWOOD PLACE} \code{796 OAKDALE AVENUE} \code{805 OAKRIDGE LANE} \code{807 JEROME AVENUE} \code{811 OLEN THOMAS DRIVE} \code{820 BUCKHANNAN AVENUE} \code{825 LINDEN AVENUE} \code{827 SCOTT STREET} \code{829 REDBUD DRIVE} \code{832 FOREST AVENUE} \code{833 PARKVIEW DRIVE} \code{837 LEO STREET} \code{838 HARLEY BROOK LANE} \code{838 VALLEY DRIVE} \code{839 OLEN THOMAS DRIVE} \code{841 HOLT STREET} \code{85 WALT NUZUM FARM ROAD} \code{851 HILL HAVEN DRIVE} \code{86 MICHAEL STREET} \code{86 WEST FORK DRIVE} \code{863 LAUREL LANE} \code{867 BROOKE STREET} \code{875 STAR TREK DRIVE} \code{886 MICHIGAN AVENUE} \code{893 JEROME AVENUE} \code{895 SIMPSON AVENUE} \code{896 MONROE AVENUE} \code{912 PHEASANT RIDGE ROAD} \code{914 SMITHFIELD AVENUE} \code{922 KENWOOD PLACE} \code{95 AGRICULTURE LANE} \code{965 FORMULA LANE} \code{966 ETHELS LANE} \code{967 FANCHER DRIVE} \code{97 CLARK STREET} \code{973 SPRUCE DRIVE} \code{98 BRANNON AVENUE} \code{988 SMALL STREET} \code{989 GIRAFFE HILL DRIVE} \code{992 FANCHER DRIVE} \code{995 HARDESTY STREET}}
    \item{\code{city}}{a factor with levels \code{ABILENE} \code{ACTON} \code{ALBANY} \code{ALLEN} \code{ALLENTOWN} \code{ALTAMONTE SPRINGS} \code{ALTOONA} \code{AMARILLO} \code{AMBLER} \code{AMENIA} \code{AMSTERDAM} \code{ANDES} \code{ARLINGTON} \code{ATHENS} \code{AUSTIN} \code{AVON PARK} \code{BABYLON} \code{BALCONES} \code{BALDWINSVILLE} \code{BARRYVILLE} \code{BARTOW} \code{BAY CITY} \code{BAYOU} \code{BEAUMONT} \code{BEAVER FALLS} \code{BEDFORD VILLAGE} \code{BEEVILLE} \code{BELLEROSE} \code{BELLEVILLE} \code{BENSALEM} \code{BIG PINE KEY} \code{BIG SPRING} \code{BINGHAMTON} \code{BLACK RIVER} \code{BLOOMING GROVE} \code{BLOOMSBURG} \code{BLOUNTSTOWN} \code{BLUE MOUND} \code{BLUFFDALE} \code{BOCA RATON} \code{BOONSVILLE} \code{BOVINA} \code{BOWLING GREEN} \code{BOYS RANCH} \code{BRADDOCK} \code{BREEZEWOOD} \code{BRENTWOOD} \code{BRIDGE CITY} \code{BRIDGEVILLE} \code{BRONSON} \code{BRONX} \code{BROOKLYN} \code{BROOKSVILLE} \code{BROWNSVILLE} \code{BRYAN} \code{BUFFALO} \code{BULLARD} \code{BUTLER} \code{CANANDAIGUA} \code{CANONSBURG} \code{CARLISLE} \code{CARRIZO SPRINGS} \code{CARROLLTON} \code{CAYUGA} \code{CHEEKTOWAGA} \code{CHESTER} \code{CLARION} \code{CLEARWATER} \code{CLEVELAND} \code{CLIFFORD} \code{COCOA} \code{COHOCTON} \code{COLONIE} \code{COMMACK} \code{CONROE} \code{CONSHOHOCKEN} \code{CORNING} \code{CORONA} \code{CORPUS CHRISTI} \code{CORTLAND} \code{COUPLAND} \code{CRAFTON} \code{CRIDERS CORNERS} \code{CUDDEBACKVILLE} \code{CYPRESS GARDENS} \code{CYPRESS LAKE} \code{DALHART} \code{DALLAS} \code{DAYTONA BEACH} \code{DECATUR} \code{DEER PARK} \code{DEERFIELD BEACH} \code{DEERPARK} \code{DELRAY BEACH} \code{DELTONA} \code{DENISON} \code{DENTON} \code{DICKINSON} \code{DOBBS FERRY} \code{DORAL} \code{DUKE CENTER} \code{DUNDEE} \code{EAGLEVILLE} \code{EAST GLENVILLE} \code{EAST PETERSBURG} \code{EAST ROCHESTER} \code{EDWARDS} \code{EL CAMPO} \code{EL PASO} \code{ELMHURST} \code{ELMSFORD} \code{EPHRATA} \code{ERIE} \code{EVERETT} \code{EVERGREEN} \code{EXTON} \code{FAIRFIELD} \code{FALLSBURGE} \code{FANNETT} \code{FARMERS BRANCH} \code{FARMINGDALE} \code{FEATHER SOUND} \code{FERNANDINA BEACH} \code{FISHERTOWN} \code{FLEISCHMANNS} \code{FLETCHER CARTER} \code{FLOMOT} \code{FOREST HILLS} \code{FORT LAUDERDALE} \code{FORT MYERS} \code{FORT PIERCE} \code{FORT WALTON BEACH} \code{FORT WASHINGTON} \code{FORT WORTH} \code{FREDERICKSBURG} \code{FRIO CANYON} \code{FRISCO} \code{FT LAUDERDALE} \code{GAINESVILLE} \code{GALVESTON} \code{GARDEN CITY} \code{GEORGETOWN} \code{GETTYSBURG} \code{GIBSONIA} \code{GLEN ROSE} \code{GOLDEN GATE (COLLIER)} \code{GRAND ISLAND} \code{GRAND PRAIRIE} \code{GREEN COVE SPRINGS} \code{H966STEAD} \code{HAMSHIRE} \code{HARLINGEN} \code{HARRISBURG} \code{HARRISON} \code{HARVEYS LAKE} \code{HATBORO} \code{HENRIETTA} \code{HEWITT} \code{HIALEAH} \code{HICKSVILLE} \code{HITCHCOCK} \code{HOLLEY} \code{HOLLEY NAVARRE} \code{HOLLYWOOD} \code{HOMESTEAD} \code{HOMOSASSA SPRINGS} \code{HONEOYE FALLS} \code{HOOVERSVILLE} \code{HORSHAM} \code{HOUSTON} \code{HUDSON} \code{HUMMELSTOWN} \code{HUNTINGTON} \code{HUXLEY} \code{ICKESBURG} \code{INDIANA} \code{IRVING} \code{IRWIN} \code{ITHACA} \code{JACKSONVILLE} \code{JACKSONVILLE BEACH} \code{JAYTON} \code{JOHNSTOWN} \code{JONESTOWN (LEBANON)} \code{JUPITER} \code{KATY} \code{KELLER} \code{KERENS} \code{KILLEEN} \code{KING OF PRUSSIA} \code{KINGSTON} \code{KISSIMMEE} \code{LA GRANGE (DUTCHESS)} \code{LA PORTE} \code{LAKE BUENA VISTA} \code{LAKE HELEN} \code{LAKE KATRINE} \code{LAKE WORTH} \code{LAKELAND} \code{LANGHORNE} \code{LANSDALE} \code{LAREDO} \code{LEAGUE CITY} \code{LEBANON} \code{LEESBURG} \code{LEVITTOWN} \code{LIVONIA} \code{LLANO} \code{LOCKHART} \code{LOGANTON} \code{LONG ISLAND CITY} \code{LONGVIEW} \code{LONGWOOD} \code{LOYALSOCK} \code{LUBBOCK} \code{M966HIS} \code{MACCLENNY} \code{MADISON} \code{MAITLAND} \code{MANHASSET} \code{MANHATTAN} \code{MANSFIELD} \code{MARIANNA} \code{MARIETTA} \code{MARLIN} \code{MASONVILLE} \code{MASSENA} \code{MAYVILLE} \code{MC KINNEY} \code{MCADOO} \code{MCALLEN} \code{MCKEANSBURG} \code{MCMURRAY} \code{MELBOURNE} \code{MERCURY} \code{MERIT} \code{MEXIA} \code{MIAMI} \code{MIAMI SPRINGS} \code{MIDLAND} \code{MIDLANDS} \code{MILTON} \code{MINEOLA} \code{MIRAMAR} \code{MOIRA} \code{MONESSEN} \code{MORAVIA} \code{MORRIS} \code{MOUNT PLEASANT MILLS} \code{MOUNT VERNON} \code{MOUNTVILLE} \code{MULLIN} \code{NACOGDOCHES} \code{NAVASOTA} \code{NAZARETH} \code{NEDERLAND} \code{NEW BRAUNFELS} \code{NEW CASTLE} \code{NEW HYDE PARK} \code{NEW PORT RICHEY} \code{NEW ROCHELLE} \code{NEW YORK} \code{NEWARK} \code{NEWBURGH} \code{NEWCOMB} \code{NEWTOWN} \code{NIAGARA FALLS} \code{NOCONA} \code{NORRISTOWN} \code{NORTH CLOVE} \code{NORTH DADE} \code{NORTH GREENBUSH} \code{NORTH NAPLES} \code{NORTHAMPTON} \code{OCALA} \code{ODESSA} \code{OKLAWAHA} \code{OLEAN} \code{ONEONTA} \code{OPA LOCKA} \code{ORANGEBURG} \code{ORLANDO} \code{ORMOND BEACH} \code{OVIEDO} \code{PAINT ROCK} \code{PALATKA} \code{PALESTINE} \code{PALM BAY} \code{PANAMA CITY} \code{PAOLI} \code{PEACOCK} \code{PEARL RIVER} \code{PECOS} \code{PEEKSKILL} \code{PENSACOLA} \code{PHILA} \code{PHILADELPHIA} \code{PINEHURST} \code{PIPE CREEK} \code{PITTSBURGH} \code{PITTSTON} \code{PLANO} \code{PLANT CITY} \code{PLANTATION} \code{PLATTSBURGH} \code{PLEASANTVILLE} \code{PLYMOUTH MEETING} \code{POLK CITY} \code{POMONA} \code{POMONA PARK} \code{POMPANO BEACH} \code{PORT O CONNOR} \code{PORT ORANGE} \code{PORT ST JOE} \code{PORT ST LUCIE} \code{PORTLAND} \code{POTTSVILLE} \code{POUGHKEEPSIE} \code{PREMONT} \code{QUAKERTOWN} \code{RED OAK} \code{REGO PARK QUEENS} \code{RICHARDSON} \code{RICHLAND} \code{RIDLEY PARK} \code{ROCHESTER} \code{ROCKWALL} \code{ROME} \code{ROSSTON} \code{ROWLETT} \code{SAN ANGELO} \code{SAN ANTONIO} \code{SAN AUGUSTINE} \code{SANFORD} \code{SARANAC LAKE} \code{SARASOTA} \code{SARATOGA SPRINGS} \code{SCHENECTADY} \code{SCOTCHTOWN} \code{SCRANTON} \code{SEAGROVE BEACH} \code{SEVEN SPRINGS} \code{SEWICKLEY} \code{SHENANDOAH} \code{SHERBURNE} \code{SHERMAN} \code{SHIPPENSBURG} \code{SINKING SPRING} \code{SINTON} \code{SKANEATELES} \code{SLIPPERY ROCK} \code{SMITHSONS VALLEY} \code{SOUDERTON} \code{SOUTH OTSELIC} \code{SPEARMAN} \code{SPRING} \code{STAFFORD} \code{STATE COLLEGE} \code{STATEN ISLAND} \code{STRASBURG} \code{STROUDSBURG} \code{STUART} \code{SUGAR LAND} \code{SUGAR NOTCH} \code{SYOSSET} \code{SYRACUSE} \code{TALLAHASSEE} \code{TAMPA} \code{TEXARKANA} \code{TEXAS CITY} \code{THOMPSON} \code{TOMBALL} \code{TOWANDA} \code{TROY} \code{TURBOTVILLE} \code{TYLER} \code{UMBARGER} \code{UTICA} \code{UVALDE} \code{VALLEY VIEW} \code{VENICE} \code{VICTORIA} \code{WACO} \code{WARMINSTER} \code{WASHINGTON} \code{WAXAHACHIE} \code{WAYNE} \code{WEBSTER} \code{WEDGWOOD} \code{WEST BRUNSWICK TWP} \code{WEST NYACK} \code{WEST PALM BEACH} \code{WEST SENECA} \code{WESTBURY} \code{WHITE HAVEN} \code{WHITE PLAINS} \code{WHITESTONE} \code{WICHITA FALLS} \code{WILDWOOD} \code{WILKES BARRE} \code{WILKINSBURG} \code{WILLIS} \code{WINDSOR} \code{WINTER HAVEN} \code{WINTER PARK} \code{WOODMERE} \code{WORTHINGTON} \code{YOAKUM} \code{YONKERS} \code{YOUNGSTOWN}}
    \item{\code{statprov}}{a factor with levels \code{FL} \code{NY} \code{PA} \code{TX}}
    \item{\code{postal}}{a numeric vector}
    \item{\code{country}}{a factor with levels \code{US}}
  }
}
\source{
Excerpt from list of EPA superfund sites 
http://www.epa.gov/superfund/sites/}
\examples{
#load("addresses.rda")
#addrListLookup(addresses)
}
\keyword{datasets}
