\name{baseGeoCoded}
\alias{baseGeoCoded}
\docType{data}
\title{
	This is the geocoded EPA List.
}
\description{
	This is the EPA List after applying the geoCode Function
}
\usage{data(baseGeoCoded)}
\format{
  A data frame with 1146 observations on the following 9 variables.
  \describe{
    \item{\code{id}}{id field}
    \item{\code{address}}{Address Data}
    \item{\code{city}}{City Datai}
    \item{\code{stateprov}}{State or Province}
    \item{\code{postal}}{Postal Code}
    \item{\code{geoURL}}{geoURL Data}
    \item{\code{geoLL}}{Raw Latitude and Longitude}
    \item{\code{long}}{split longitude}
    \item{\code{lat}}{split latitude}
  }
}
\details{}
\source{}
\references{}
\examples{}
\keyword{datasets}
