\name{baseParsed}
\alias{baseParsed}
\docType{data}
\title{Parsed Base Data Set}
\description{
	This is the 'EPA List.csv' file parsed by address.
}
\usage{data(baseParsed)}
\format{
  A data frame with the following variables.
  \describe{
    \item{\code{id}}{a unique key field for the user's use}
    \item{\code{address}}{Address data}
    \item{\code{city}}{City Data}
    \item{\code{stateprov}}{State or Province}
    \item{\code{postal}}{Postal Code}
    \item{\code{addNums}}{parsed match field for concatenation algorithm}
  }
}
\details{}
\source{
}
\references{}
\examples{}
\keyword{datasets}
