\name{geoComplete}
\alias{geoComplete}
\docType{data}
\title{Complete Geocoded Comparison File}
\description{
		This is the completed compared geocoded results list.
}
\usage{data(geoComplete)}
\format{
  A data frame with the following variables.
  \describe{
    \item{\code{shortKey}}{short id field}
    \item{\code{Match_id}}{self explanatory}
    \item{\code{Match_address}}{self explanatory}
    \item{\code{Match_city}}{self explanatory}
    \item{\code{Match_stateprov}}{self explanatory}
    \item{\code{Match_postal}}{self explanatory}
    \item{\code{Match_geoURL}}{self explanatory}
    \item{\code{Match_geoLL}}{self explanatory}
    \item{\code{Match_longitude}}{self explanatory}
    \item{\code{Match_latitude}}{self explanatory}
    \item{\code{Base_id}}{self explanatory}
    \item{\code{Base_address}}{self explanatory}
    \item{\code{Base_city}}{self explanatory}
    \item{\code{Base_stateprov}}{self explanatory}
    \item{\code{Base_postal}}{self explanatory}
    \item{\code{Base_geoURL}}{self explanatory}
    \item{\code{Base_geoLL}}{self explanatory}
    \item{\code{Base_longitude}}{self explanatory}
    \item{\code{Base_latitude}}{self explanatory}
    \item{\code{geoValid}}{self explanatory}
    \item{\code{distanceKm}}{self explanatory}
    \item{\code{distanceMi}}{self explanatory}
  }
}
\details{}
\source{}
\references{}
\examples{}
\keyword{datasets}
