\name{ips}
\alias{ips}
\docType{data}
\title{ips/data}
\description{Sample ips data for geocoding}
\usage{data(ips)}
\format{
  A data frame with 250 observations on the following variable.
  \describe{
    \item{\code{ip}}{a factor with levels \code{108.128.64.50} \code{108.223.56.106} \code{108.230.22.18} \code{108.231.204.187} \code{108.66.55.227} \code{108.78.203.83} \code{108.87.12.8} \code{108.89.242.168} \code{108.93.16.147} \code{109.112.215.78} \code{109.13.101.13} \code{109.148.136.113} \code{109.157.0.65} \code{109.202.94.78} \code{110.33.130.123} \code{113.210.97.65} \code{114.44.55.232} \code{114.79.1.31} \code{115.132.63.70} \code{117.198.59.246} \code{118.101.15.85} \code{118.101.170.23} \code{119.154.233.101} \code{120.22.188.231} \code{121.217.172.75} \code{121.72.81.200} \code{121.96.13.251} \code{122.63.75.54} \code{124.155.89.235} \code{14.207.207.83} \code{141.0.8.83} \code{151.48.116.162} \code{164.77.53.246} \code{173.169.138.123} \code{174.108.34.41} \code{174.131.138.47} \code{174.28.70.123} \code{176.27.103.251} \code{178.98.148.57} \code{180.190.249.54} \code{180.93.253.68} \code{182.63.6.197} \code{184.13.159.13} \code{184.32.206.179} \code{184.36.52.126} \code{184.39.60.31} \code{184.42.7.47} \code{184.45.6.171} \code{184.46.11.77} \code{184.47.160.219} \code{184.90.167.244} \code{186.105.128.98} \code{186.20.37.22} \code{187.195.127.82} \code{188.28.47.231} \code{188.7.34.179} \code{189.102.225.156} \code{189.143.128.12} \code{189.93.168.85} \code{190.14.206.242} \code{190.160.127.61} \code{190.164.217.21} \code{190.204.43.115} \code{190.246.142.227} \code{190.255.251.58} \code{190.96.187.54} \code{193.250.231.184} \code{2.49.245.73} \code{200.112.106.61} \code{200.112.239.35} \code{200.35.216.151} \code{200.71.151.210} \code{201.165.20.178} \code{208.42.212.2} \code{208.54.40.247} \code{208.54.64.244} \code{208.61.65.161} \code{213.122.105.162} \code{213.87.142.214} \code{217.233.137.139} \code{217.39.63.197} \code{217.55.75.69} \code{24.129.76.176} \code{24.139.225.169} \code{24.16.186.202} \code{24.190.196.108} \code{27.51.11.175} \code{27.7.139.114} \code{31.108.38.68} \code{37.24.208.203} \code{41.105.126.21} \code{41.234.80.212} \code{46.60.252.96} \code{50.42.233.2} \code{50.51.209.209} \code{50.51.3.158} \code{59.177.79.50} \code{62.203.244.38} \code{62.31.147.145} \code{62.31.147.67} \code{62.62.239.60} \code{64.124.189.104} \code{65.1.162.223} \code{65.129.167.177} \code{65.5.222.113} \code{66.156.233.51} \code{66.157.57.254} \code{67.140.194.182} \code{67.149.7.35} \code{67.52.126.117} \code{68.126.186.162} \code{68.253.225.201} \code{68.34.55.199} \code{68.63.11.175} \code{69.113.225.209} \code{69.152.227.183} \code{69.225.25.211} \code{69.237.117.106} \code{70.145.87.104} \code{70.165.202.187} \code{70.2.53.206} \code{70.234.231.193} \code{70.246.146.209} \code{70.251.64.75} \code{70.46.105.202} \code{71.42.81.153} \code{71.67.107.8} \code{72.145.158.40} \code{72.147.48.95} \code{72.152.43.202} \code{72.152.45.65} \code{72.152.67.220} \code{74.102.148.91} \code{74.138.101.23} \code{74.141.226.83} \code{74.161.0.240} \code{74.161.49.78} \code{74.176.153.192} \code{74.176.30.247} \code{74.178.204.201} \code{74.179.182.223} \code{74.179.53.215} \code{74.190.116.2} \code{74.233.11.243} \code{74.233.198.125} \code{74.234.163.249} \code{74.235.107.144} \code{74.235.139.35} \code{74.235.228.176} \code{74.240.229.42} \code{74.241.95.17} \code{74.243.154.37} \code{74.248.226.248} \code{74.250.167.235} \code{75.134.90.214} \code{75.205.242.104} \code{75.23.33.72} \code{75.34.49.175} \code{75.4.139.101} \code{75.4.192.4} \code{75.60.5.41} \code{76.112.50.87} \code{76.183.231.13} \code{76.19.14.51} \code{76.201.153.92} \code{76.202.137.192} \code{76.212.14.39} \code{76.226.165.111} \code{76.244.153.185} \code{77.102.168.53} \code{78.144.120.32} \code{78.144.37.179} \code{78.3.229.90} \code{79.32.91.34} \code{79.54.210.71} \code{79.88.174.157} \code{81.149.208.71} \code{82.128.59.158} \code{82.13.36.149} \code{82.239.225.191} \code{82.240.117.48} \code{83.25.52.245} \code{83.60.74.208} \code{85.182.41.81} \code{86.124.248.219} \code{86.130.203.204} \code{86.145.211.62} \code{86.150.108.177} \code{86.152.14.35} \code{86.176.181.92} \code{86.178.127.251} \code{87.18.187.17} \code{88.109.76.46} \code{89.241.60.186} \code{89.242.217.233} \code{90.16.82.166} \code{90.202.19.174} \code{90.216.128.186} \code{90.217.56.62} \code{91.177.169.114} \code{92.143.29.186} \code{92.152.79.146} \code{92.161.118.70} \code{92.22.167.35} \code{92.23.120.138} \code{92.30.89.190} \code{92.40.253.127} \code{92.60.106.210} \code{94.120.154.62} \code{94.15.76.93} \code{94.169.76.132} \code{94.175.125.2} \code{94.76.130.229} \code{95.144.198.150} \code{95.225.152.221} \code{96.24.172.173} \code{96.37.140.50} \code{98.64.91.139} \code{98.67.115.103} \code{98.68.155.53} \code{98.68.206.250} \code{98.68.209.79} \code{98.70.154.127} \code{98.71.7.102} \code{98.80.100.61} \code{98.80.238.218} \code{98.81.165.128} \code{98.83.23.76} \code{98.83.73.15} \code{98.84.144.145} \code{98.85.194.79} \code{98.88.125.182} \code{98.89.139.132} \code{98.91.51.135} \code{98.92.212.66} \code{99.0.34.2} \code{99.102.69.143} \code{99.107.178.144} \code{99.110.52.59} \code{99.111.247.114} \code{99.127.73.154} \code{99.130.196.222} \code{99.141.144.32} \code{99.152.206.214} \code{99.155.153.210} \code{99.156.195.214} \code{99.16.44.233} \code{99.18.147.45} \code{99.28.173.230} \code{99.28.255.39}}
  }
}
\source{
Randomly generated ip addresses
}
\examples{
#load("ips.rda")
#ipListLookup(ips)
}
\keyword{datasets}
