\name{matchData}
\alias{matchData}
\docType{data}
\title{Loaded Vendors.xml list}
\description{
	This is the Vendors.xml file loaded with xmlImport
}
\usage{data(matchData)}
\format{
  A data frame with 35 observations on the following 5 variables.
  \describe{
    \item{\code{id}}{a character vector}
    \item{\code{address}}{a character vector}
    \item{\code{city}}{a character vector}
    \item{\code{stateprov}}{a character vector}
    \item{\code{postal}}{a character vector}
  }
}
\details{}
\source{}
\references{}
\examples{}
\keyword{datasets}
