\name{matchGeoCoded}
\alias{matchGeoCoded}
\docType{data}
\title{
	This is the geocoded Vendor List.
}
\description{
	This is the Vendor List after applying the geoCode Function
}
\usage{data(matchGeoCoded)}
\format{
  A data frame with the following variables.
  \describe{
    \item{\code{id}}{id field}
    \item{\code{address}}{Address Data}
    \item{\code{city}}{City Datai}
    \item{\code{stateprov}}{State or Province}
    \item{\code{postal}}{Postal Code}
    \item{\code{geoURL}}{geoURL Data}
    \item{\code{geoLL}}{Raw Latitude and Longitude}
    \item{\code{long}}{split longitude}
    \item{\code{lat}}{split latitude}
  }
}
\details{}
\source{}
\references{}
\examples{}
\keyword{datasets}
