\name{shortConv}
\alias{shortConv}
\title{Truncates and Converts latitude and longitude to a whole degree}
\description{
	This function truncates and converts latitude and longitude to a whole 
	degree for comparison.  NOTE:  This is a support function and is not intended
	to be used stand alone and is a component of the encodeLLshort function.
}
\usage{
shortConv(x)
}
\arguments{
  \item{x}{x is the data unit to be converted}
}
\details{
	NOTE:  This is a support function and is not intended
	to be used stand alone.
}
\value{
	The returned value is a shortened and converted to a character value
}
\references{}
\author{
  Randall Shane, PhD \email{rshane@basexvi.com}
}
\note{}
\examples{
## NOTE:  This is a support function and is not intended to be used
##	stand alone and is a component of the encodeLLshort function.
}

\keyword{convert}

