\name{degrees2radians}
\alias{degrees2radians}
\title{degrees2radians}
\description{This is a conversion from degrees to radians for the haversine function.  This is not intended to be used independently.}
\usage{degrees2radians(x)}
\arguments{
  \item{x}{Degrees measure.}
}
\author{
Randall Shane, PhD\cr
rshane@basexvi.com
}
\keyword{haversine}
\keyword{radians}
